[MV] move maintain_attrs() from document_vars.py to tools.py
